[VACGOM-69] refactor: 자바 엔티티 제거 및 코틀린 엔티티 변경에 따른 의존 코드 수정
