--- src/main/java/kr/co/vacgom/api/user/enums/Gender.kt
@@ -1,5 +1,0 @@
-package kr.co.vacgom.api.user.domain.enums
-
-enum class Gender {
-    MALE, FEMALE
-}
