+++ src/main/kotlin/kr/co/vacgom/api/baby/repository/BabyJpaRepository.kt
@@ -1,6 +1,6 @@
 package kr.co.vacgom.api.baby.repository
 
-import kr.co.vacgom.api.baby.Baby
+import kr.co.vacgom.api.baby.domain.Baby
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface BabyJpaRepository: JpaRepository<Baby, Long>
