+++ src/main/kotlin/kr/co/vacgom/api/user/application/AuthService.kt
@@ -3,7 +3,9 @@ package kr.co.vacgom.api.user.application
 import kr.co.vacgom.api.auth.oauth.OAuthHandler
 import kr.co.vacgom.api.auth.oauth.enums.SocialLoginProvider
 import kr.co.vacgom.api.auth.security.util.SecurityContextUtil
+import kr.co.vacgom.api.global.exception.error.BusinessException
 import kr.co.vacgom.api.user.domain.User
+import kr.co.vacgom.api.user.exception.UserError
 import kr.co.vacgom.api.user.presentation.dto.Login
 import kr.co.vacgom.api.user.repository.UserRepository
 import org.springframework.stereotype.Service
@@ -36,7 +38,7 @@ class AuthService(
     }
 
     fun unlinkUser(user: User) {
-        val socialId = user.socialId
+        val socialId = user.socialId ?: throw BusinessException(UserError.SOCIAL_ID_NOT_FOUND)
         oauthHandler.handleUnlinkUser(user.provider, socialId)
     }
 }
